+++ src/assets/scripts/client/parsers/aircraftCommandParser/aircraftCommandMap.js
@@ -60,7 +60,7 @@ export const AIRCRAFT_COMMAND_MAP = {
         isSystemCommand: false
     },
     cross: {
-        aliases: ['cross', 'cr'],
+        aliases: ['cross', 'cr', 'x'],
         functionName: 'runCross',
         isSystemCommand: false
     },
@@ -231,7 +231,7 @@ export const AIRCRAFT_COMMAND_MAP = {
  * @type {array}
  * @final
  */
-export const EXPEDITE = ['expedite', 'x'];
+export const EXPEDITE = ['expedite', 'ex'];
 
 /**
  * Get the name of a command when given any of that command's aliases