+++ src/assets/scripts/client/ui/SettingsController.js
@@ -10,8 +10,8 @@ import { SELECTORS } from '../constants/selectors';
  * @final
  */
 const UI_SETTINGS_MODAL_TEMPLATE = `
-    <div class="option-dialog dialog">
-        <p class="dialog-title">Options</p>
+    <div class="option-dialog dialog notSelectable">
+        <p class="dialog-title">Settings</p>
         <div class="dialog-body nice-scrollbar"></div>
     </div>`;