+++ CHANGELOG.md
@@ -1,9 +1,67 @@
-# 6.18.0 (June 1, 2020)
+# 6.22.0 (December 1, 2020)
 ### New Features
 
 ### Bugfixes
+- <a href="https://github.com/openscope/openscope/issues/1565" target="_blank">#1565</a> - Prevent aircraft from taxiing to a runway they're already holding short of
+- <a href="https://github.com/openscope/openscope/issues/1190" target="_blank">#1190</a> - Prevent selection of text in most dialogs
 
 ### Enhancements & Refactors
+- <a href="https://github.com/openscope/openscope/issues/1614" target="_blank">#1614</a> - Spell out squawk code in single digit form in readbacks
+
+
+# 6.21.0 (November 4, 2020)
+### New Features
+- <a href="https://github.com/openscope/openscope/issues/1654" target="_blank">#1654</a> - Add Greensboro - Piedmont Triad Airport (KGSO)
+
+### Enhancements & Refactors
+- <a href="https://github.com/openscope/openscope/issues/1652" target="_blank">#1652</a> - Update EIDW/EICK airspace and procedures
+- <a href="https://github.com/openscope/openscope/issues/1658" target="_blank">#1658</a> - Update LIPZ airspace
+- <a href="https://github.com/openscope/openscope/issues/1633" target="_blank">#1633</a> - Add missing important fixes to KPDX
+- <a href="https://github.com/openscope/openscope/issues/1651" target="_blank">#1651</a> - Deactivate or fix airports with broken airspace
+
+
+# 6.20.0 (October 7, 2020)
+### Enhancements & Refactors
+- <a href="https://github.com/openscope/openscope/issues/1615" target="_blank">#1615</a> - Fix broken link in KSTL airport guide
+- <a href="https://github.com/openscope/openscope/issues/1583" target="_blank">#1583</a> - Update EDDF to AIRAC 2009
+- <a href="https://github.com/openscope/openscope/issues/1645" target="_blank">#1645</a> - Change expedite command from "x" to "ex" (since "x" is now used for "cross")
+- <a href="https://github.com/openscope/openscope/issues/391" target="_blank">#391</a> - Add support for multiple airspace sections
+
+
+# 6.19.0 (August 10, 2020)
+### New Features
+- <a href="https://github.com/openscope/openscope/issues/1624" target="_blank">#1624</a> - Add speed capability to crossing restrictions
+- <a href="https://github.com/openscope/openscope/issues/1610" target="_blank">#1610</a> - Add Sendai Airport (RJSS)
+
+### Bugfixes
+- <a href="https://github.com/openscope/openscope/issues/1617" target="_blank">#1617</a> - Add LXJ to fix crash at KIAD
+- <a href="https://github.com/openscope/openscope/issues/1626" target="_blank">#1626</a> - Draw range rings where the file specifies instead of at the airport center
+
+### Enhancements & Refactors
+- <a href="https://github.com/openscope/openscope/issues/1628" target="_blank">#1628</a> - Use Github status checks for all pull requests
+- <a href="https://github.com/openscope/openscope/issues/1538" target="_blank">#1538</a> - Update KDCA
+
+
+# 6.18.0 (July 1, 2020)
+### New Features
+- <a href="https://github.com/openscope/openscope/issues/1523" target="_blank">#1523</a> - Re-introduce Washington Dulles International Airport (KIAD); add Mountain Air Cargo (MTN) and CommutAir (UCA)
+- <a href="https://github.com/openscope/openscope/issues/1568" target="_blank">#1568</a> - Add command reference and slack links to help menu
+
+### Bugfixes
+- <a href="https://github.com/openscope/openscope/issues/39" target="_blank">#39</a> - Fix misaligned compass headings around scope border
+
+### Enhancements & Refactors
+- <a href="https://github.com/openscope/openscope/issues/1590" target="_blank">#1590</a> - Update KLAX airport guide description
+- <a href="https://github.com/openscope/openscope/issues/1589" target="_blank">#1589</a> - Update KDCA airport guide links
+- <a href="https://github.com/openscope/openscope/issues/1581" target="_blank">#1581</a> - Add ability to reset scratchpad with "."
+- <a href="https://github.com/openscope/openscope/issues/1606" target="_blank">#1606</a> - Retitle options menu to "Settings"
+- <a href="https://github.com/openscope/openscope/issues/1608" target="_blank">#1608</a> - Add/Update Japanese Airline Files
+
+
+# 6.17.1 (May 21, 2020)
+### Hotfixes
+- <a href="https://github.com/openscope/openscope/issues/1573" target="_blank">#1573</a> - Add missing waypoints from KJFK to fix crash
+
 
 # 6.17.0 (May 1, 2020)
 ### New Features
@@ -16,6 +74,7 @@
 - <a href="https://github.com/openscope/openscope/issues/1543" target="_blank">#1543</a> - Remove BEE/TCX airlines
 - <a href="https://github.com/openscope/openscope/issues/1551" target="_blank">#1551</a> - Add missing waypoint CORDS at KJFK
 - <a href="https://github.com/openscope/openscope/issues/1271" target="_blank">#1271</a> - Update the tutorial
+
 
 # 6.16.0 (March 1, 2020)
 ### New Features